split common helpers into files and add span helpers
Break tests/common/mod.rs into exec, asserts, spans, and macros modules,
re-exporting everything from mod.rs so existing call sites are unchanged.

Add span helpers that derive Span from the source instead of hardcoded
byte offsets: span_of, span_of_last, span_of_nth, span_whole, span_up_to.
Each panics with a descriptive message when a needle is absent, and the
helpers carry their own unit tests.
